Decorate pattern matches somewhat
