+++ wasm-calc10/src/Calc/Linearity/Types.hs
@@ -8,16 +8,21 @@ module Calc.Linearity.Types
     LinearityType (..),
     LinearState (..),
     UserDefined (..),
+    Drops(..)
   )
 where
 
+import Calc.Types.Type
+import qualified Data.List.NonEmpty as NE
 import Calc.Types.Identifier
 import qualified Data.Map as M
 import GHC.Natural
 
--- | Are we using the whole type or bits of it?
--- this distinction will be gone once we can destructure types instead,
--- implicitly destroying them
+data Drops ann
+  = DropIdentifiers (NE.NonEmpty (Identifier, Type ann))
+  | DropMe
+  deriving stock (Eq, Ord, Show, Functor)
+
 newtype Linearity ann
   = Whole ann
   deriving newtype (Eq, Ord, Show)
