+++ wasm-calc10/src/Calc/Typecheck/Error.hs
@@ -1,5 +1,8 @@
-module Calc.Typecheck.Error (module Calc.Typecheck.Error.TypeError,
-    module Calc.Typecheck.Error.PatternMatchError) where
+module Calc.Typecheck.Error
+  ( module Calc.Typecheck.Error.TypeError,
+    module Calc.Typecheck.Error.PatternMatchError,
+  )
+where
 
-import Calc.Typecheck.Error.TypeError
 import Calc.Typecheck.Error.PatternMatchError
+import Calc.Typecheck.Error.TypeError
