+++ wasm-calc10/test/Test/Typecheck/TypecheckSpec.hs
@@ -177,9 +177,9 @@ spec = do
                   ],
                 tyInt32
               ),
-              ("function main() -> Int32 { case True { True -> 1, False -> 2 } }", 
-                  tyInt32)
-
+              ( "function main() -> Int32 { case True { True -> 1, False -> 2 } }",
+                tyInt32
+              )
             ]
       describe "Successfully typechecking modules" $ do
         traverse_ testSucceedingModule succeeding
